cloud_function/requirements.txt
functions-framework==3.8.2
google-api-python-client==2.163.0
google-auth==2.38.0
